Add moderation replies
Allows a reply to be included with each moderation which is visible in
the history table on the moderation page, and to the experience author
on the edit story page.

In this implementation, moderators can see all previous replies in the
table, but the author can see only the most recent reply.