Add moderation replies
Allows a reply to be included with each moderation which is visible in
the history table on the moderation page, and to the experience author
on the edit story page.

Moderation is performed using a menu which appears when portions of the
experience text are selected. The moderator can then select specific
reasons for moderating the story from the menu, which will be passed
back to the author.