Add user profile page
Adds an initial user profile page. The profile fields are just examples
at this stage and will need refinement. The implementation adds a new
"users" app which stores profile data linked to the authenticated in
User instance.

The user is taken to the profile page directly after their first login
and encouraged to completed the details. At other times it can be
accessed from the account menu at the top right of the page. The trigger
fields are used as the default option when the search page is entered.

Co-authored-by: Bastian Greshake Tzovaras <bgreshake@googlemail.com>