Avoid help page menu overlapping text
When the width of the viewport is reduced below 992 pixels the menu
along the left hand edge overlaps the text on the page. Using Bootstraps
columns it should jump above the text, but this is prevented because the
positioning of the menu is set to fixed (so that it stays on screen when
the page is scrolled).

This change fixes this by creating a media selector that makes the menu
position relative when the width of the page reduces below 992 pixels.
An invisible column is also added to avoid the need to place the menu
inside another column, and to prevent the text from shifting to the left
when the menu position is set to fixed.

The footer position is also set to relative so that it goes over (rather
than under) the menu to avoids a nasty overlap when viewport height is
small.

Contributes to #576.